requirements.txt
aws-cdk-lib==2.88.0
constructs>=10.0.0,<11.0.0
cdk-stacksets==0.0.148
